app/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
pymongo==4.10.1
motor==3.3.2

# Web Scraping and Automation
selenium==4.18.1
beautifulsoup4==4.12.3
requests==2.31.0
webdriver-manager==4.0.1
apify-client==1.8.0

# Scheduling and Background Tasks
apscheduler==3.10.4

# Data Validation and Settings
pydantic==2.9.2
python-dotenv==1.0.1
python-decouple==3.8

# Notifications
twilio==9.3.0

# Utilities
python-dateutil==2.8.2
aiohttp==3.9.3
httpx==0.27.0 